+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/shader/CobblemonShaders.kt
@@ -25,10 +25,10 @@ object CobblemonShaders {
         SHADERS_TO_REGISTER.add(Pair(shader, callback))
     }
     fun init(){
-        registerShader({rm: ResourceProvider -> ShaderRegistryData(rm, cobblemonResource("particle_add"), DefaultVertexFormat.POSITION_COLOR_TEX_LIGHTMAP) }) {
+        registerShader({rm: ResourceProvider -> ShaderRegistryData(rm, cobblemonResource("particle_add"), DefaultVertexFormat.PARTICLE) }) {
             PARTICLE_BLEND = it
         }
-        registerShader({rm: ResourceProvider -> ShaderRegistryData(rm, cobblemonResource("particle_cutout"), DefaultVertexFormat.POSITION_COLOR_TEX_LIGHTMAP) }) {
+        registerShader({rm: ResourceProvider -> ShaderRegistryData(rm, cobblemonResource("particle_cutout"), DefaultVertexFormat.PARTICLE) }) {
             PARTICLE_CUTOUT = it
         }
     }